# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NEAR Proof Arena API",
    "version": "0.1.0",
    "description": "Public API (default port 8471) and internal worker API (`/internal/v1`, separate listener, default port 8472). Contracts: docs/CONTRACTS.md. All candidate-provided strings are sanitized plain text."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8471"
    }
  ],
  "paths": {
...
    "/internal/v1/jobs/lease": {
      "post": {
        "summary": "Worker API (port 8472): lease the next job this worker may run",
        "tags": [
          "worker"
        ],
        "responses": {
          "200": {
            "description": "leased job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeasedJob"
                }
              }
            }
          },
          "204": {
            "description": "no runnable job"
          },
          "401": {
            "description": "missing/invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseRequest"
              }
            }
          }
        }
      }
    },
    "/internal/v1/jobs/{id}/heartbeat": {
      "post": {
        "summary": "Worker API: extend a lease; `cancelled=true` means stop",
        "tags": [
          "worker"
        ],
        "responses": {
          "200": {
            "description": "lease extended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          },
          "409": {
            "description": "lease lost",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "job id (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        }
      }
    },
    "/internal/v1/jobs/{id}/complete": {
      "post": {
        "summary": "Worker API: report the job result",
        "tags": [
          "worker"
        ],
        "responses": {
          "200": {
            "description": "recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "409": {
            "description": "lease lost / cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "invalid result (counted as an infra failure) or artifact not uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "job id (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          }
        }
      }
    },
    "/internal/v1/jobs/{id}/fail": {
      "post": {
        "summary": "Worker API: report an infrastructure failure (bounded retries)",
        "tags": [
          "worker"
        ],
        "responses": {
          "200": {
            "description": "recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "409": {
            "description": "lease lost / cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "job id (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailRequest"
              }
            }
          }
        }
      }
    },
    "/internal/v1/artifacts/{digest}": {
      "get": {
        "summary": "Worker API: download an object by digest",
        "tags": [
          "worker"
        ],
        "responses": {
          "200": {
            "description": "bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "digest",
            "in": "path",
            "required": true,
            "description": "sha256:<hex>",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "Worker API: upload an object; the body must hash to the path digest",
        "tags": [
          "worker"
        ],
        "responses": {
          "201": {
            "description": "stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPutResponse"
                }
              }
            }
          },
          "200": {
            "description": "already present",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPutResponse"
                }
              }
            }
          },
          "400": {
            "description": "digest mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "workerToken": []
          }
        ],
        "parameters": [
          {
            "name": "digest",
            "in": "path",
            "required": true,
            "description": "sha256:<hex>",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AckResponse": {
        "type": "object",
        "required": [
          "message",
          "ok"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ArtifactPutResponse": {
        "type": "object",
        "required": [
          "digest",
          "size_bytes"
        ],
        "properties": {
          "digest": {
            "$ref": "#/components/schemas/Digest"
          },
          "size_bytes": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0
          }
        }
      },
...
      "BenchmarkResult": {
        "type": "object",
        "required": [
          "classes",
          "hardware_profile",
          "measured_by",
          "prepare_ns",
          "public_artifact_bytes",
          "suite_revision"
        ],
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassMeasurement"
            }
          },
          "hardware_profile": {
            "type": "string"
          },
          "measured_by": {
            "type": "string"
          },
          "prepare_ns": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0
          },
          "public_artifact_bytes": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0
          },
          "score_ci_milli": {
            "description": "Half-width of a bootstrap 95% interval, milli units.",
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "score_milli": {
            "description": "Score * 1000 as integer (e.g. 100000 == 100.000).",
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "suite_revision": {
            "type": "string"
          }
        }
      },
...
      "BuildJob": {
        "type": "object",
        "required": [
          "challenge",
          "ctx",
          "manifest"
        ],
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "ctx": {
            "$ref": "#/components/schemas/JobContext"
          },
          "manifest": {
            "$ref": "#/components/schemas/CandidateManifest"
          }
        }
      },
      "BuildOutputs": {
        "description": "Outputs of the judge-run build (and judge-run `prepare`).",
        "type": "object",
        "required": [
          "bundle",
          "certificate_decl",
          "formal_tree",
          "prepare",
          "prove",
          "public_artifacts",
          "verify"
        ],
        "properties": {
          "build_ns": {
            "description": "Judge-measured wall time of one build, ns (informational).",
            "default": null,
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "bundle": {
            "description": "Tree digest of the complete built bundle (all build outputs).",
            "$ref": "#/components/schemas/Digest"
          },
          "certificate_decl": {
            "description": "Lean constant named by the manifest's `[formal].certificate` (empty if none).",
            "type": "string"
          },
          "formal_tree": {
            "description": "Tree digest of the candidate's `formal/` Lean project.",
            "$ref": "#/components/schemas/Digest"
          },
          "prepare": {
            "$ref": "#/components/schemas/Digest"
          },
          "prove": {
            "$ref": "#/components/schemas/Digest"
          },
          "public_artifacts": {
            "description": "Tree digest of `public_dir` produced by the judge-run `prepare`, frozen by digest.",
            "$ref": "#/components/schemas/Digest"
          },
          "toolchain_image": {
            "description": "Build sandbox image/rootfs digest or id (informational).",
            "default": null,
            "type": "string",
            "nullable": true
          },
          "verify": {
            "$ref": "#/components/schemas/Digest"
          }
        }
      },
...
      "CandidateManifest": {
        "description": "Parsed `candidate.toml`. Every field is a *claim* the judge validates.",
        "type": "object",
        "required": [
          "agent",
          "backend_family",
          "build",
          "challenge",
          "entry",
          "hardware",
          "name",
          "schema",
          "security_profile_request"
        ],
        "properties": {
          "agent": {
            "type": "string"
          },
          "backend_family": {
            "type": "string"
          },
          "build": {
            "$ref": "#/components/schemas/BuildSection"
          },
          "challenge": {
            "type": "string"
          },
          "entry": {
            "$ref": "#/components/schemas/EntrySection"
          },
          "formal": {
            "$ref": "#/components/schemas/FormalSection",
            "nullable": true
          },
          "hardware": {
            "$ref": "#/components/schemas/HardwareRequest"
          },
          "name": {
            "type": "string"
          },
          "parent": {
            "type": "string",
            "nullable": true
          },
          "schema": {
            "type": "string"
          },
          "security_profile_request": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CompleteRequest": {
        "type": "object",
        "required": [
          "lease_id",
          "result"
        ],
        "properties": {
          "lease_id": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/JobResult"
          }
        }
      },
...
      "Digest": {
        "description": "`sha256:<64 lowercase hex>`",
        "type": "string",
        "pattern": "^sha256:[0-9a-f]{64}$"
      },
...
      "EvidenceGraph": {
        "type": "object",
        "required": [
          "edges",
          "nodes"
        ],
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceEdge"
            }
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceNode"
            }
          }
        }
      },
      "EvidenceNode": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "label"
        ],
        "properties": {
          "digest": {
            "$ref": "#/components/schemas/Digest",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/NodeKind"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "EvidenceRef": {
        "type": "object",
        "required": [
          "digest",
          "label",
          "public"
        ],
        "properties": {
          "digest": {
            "$ref": "#/components/schemas/Digest"
          },
          "label": {
            "type": "string"
          },
          "public": {
            "description": "Whether this artifact may be shown publicly (held-out data must not).",
            "type": "boolean"
          }
        }
      },
      "ExecJob": {
        "description": "Payload shared by the post-build execution stages.",
        "type": "object",
        "required": [
          "build",
          "challenge",
          "ctx",
          "manifest"
        ],
        "properties": {
          "build": {
            "$ref": "#/components/schemas/BuildOutputs"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "ctx": {
            "$ref": "#/components/schemas/JobContext"
          },
          "manifest": {
            "$ref": "#/components/schemas/CandidateManifest"
          }
        }
      },
      "ExecutionInfo": {
        "description": "How (and with what isolation) a result was produced. Results produced by the `bwrap-dev` sandbox or by any simulated component must carry `tier_cap = demo`.",
        "type": "object",
        "required": [
          "sandbox_backend",
          "tier_cap",
          "worker_version"
        ],
        "properties": {
          "sandbox_backend": {
            "description": "e.g. `firecracker`, `bwrap-dev`.",
            "type": "string"
          },
          "tier_cap": {
            "$ref": "#/components/schemas/Tier"
          },
          "worker_version": {
            "type": "string"
          }
        }
      },
      "FailRequest": {
        "type": "object",
        "required": [
          "error",
          "lease_id",
          "retryable"
        ],
        "properties": {
          "error": {
            "description": "Infrastructure error description (sanitized and bounded by the server).",
            "type": "string"
          },
          "lease_id": {
            "type": "string"
          },
          "retryable": {
            "description": "`true` for transient errors (retried up to `max_attempts`).",
            "type": "boolean"
          }
        }
      },
      "FormalCheckJob": {
        "type": "object",
        "required": [
          "build",
          "challenge",
          "ctx",
          "manifest",
          "verified_surface"
        ],
        "properties": {
          "build": {
            "$ref": "#/components/schemas/BuildOutputs"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "ctx": {
            "$ref": "#/components/schemas/JobContext"
          },
          "manifest": {
            "$ref": "#/components/schemas/CandidateManifest"
          },
          "verified_surface": {
            "$ref": "#/components/schemas/VerifiedSurface"
          }
        }
      },
...
      "GateResult": {
        "type": "object",
        "required": [
          "evidence",
          "gate",
          "mandatory",
          "reason_codes",
          "status",
          "summary"
        ],
        "properties": {
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            }
          },
          "finished_at": {
            "type": "string",
            "nullable": true
          },
          "gate": {
            "description": "Gate id; equals the obligation id for obligation gates.",
            "$ref": "#/components/schemas/ObligationId"
          },
          "mandatory": {
            "type": "boolean"
          },
          "reason_codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            }
          },
          "reused_from": {
            "description": "Set when the result was reused from a parent via the content-addressed cache.",
            "type": "string",
            "nullable": true
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/GateStatus"
          },
          "summary": {
            "description": "Bounded, sanitized, plain text (never HTML).",
            "type": "string"
          }
        }
      },
...
      "HeartbeatRequest": {
        "type": "object",
        "required": [
          "lease_id"
        ],
        "properties": {
          "extend_seconds": {
            "default": null,
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0,
            "nullable": true
          },
          "lease_id": {
            "type": "string"
          }
        }
      },
      "HeartbeatResponse": {
        "type": "object",
        "required": [
          "cancelled",
          "lease_until"
        ],
        "properties": {
          "cancelled": {
            "description": "The run was cancelled: stop work and do not call `/complete`.",
            "type": "boolean"
          },
          "lease_until": {
            "type": "string"
          }
        }
      },
...
      "JobContext": {
        "description": "Context common to every job of a run.",
        "type": "object",
        "required": [
          "challenge_digest",
          "challenge_id",
          "package_digest",
          "run_id",
          "submission_id",
          "tier"
        ],
        "properties": {
          "challenge_digest": {
            "$ref": "#/components/schemas/Digest"
          },
          "challenge_id": {
            "type": "string"
          },
          "package_digest": {
            "description": "Digest of the uploaded candidate package archive (fetch via the artifact API).",
            "$ref": "#/components/schemas/Digest"
          },
          "run_id": {
            "type": "string"
          },
          "submission_id": {
            "type": "string"
          },
          "tier": {
            "description": "Tier of the challenge. A worker whose sandbox cannot satisfy this tier must refuse the job (the control plane never leases it to one).",
            "$ref": "#/components/schemas/Tier"
          }
        }
      },
      "JobKind": {
        "description": "Kind of a pipeline stage job.",
        "type": "string",
        "enum": [
          "VALIDATE",
          "BUILD",
          "FORMAL_CHECK",
          "CONFORMANCE",
          "ADVERSARIAL",
          "BENCHMARK"
        ]
      },
      "JobResult": {
        "description": "Everything a worker reports for one job.",
        "type": "object",
        "required": [
          "artifacts",
          "execution",
          "gates"
        ],
        "properties": {
          "artifacts": {
            "description": "Artifacts produced (logs, transcripts, exports...), already uploaded via the artifact API.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            }
          },
          "benchmark": {
            "description": "Required for `BENCHMARK` jobs that measured something.",
            "$ref": "#/components/schemas/BenchmarkResult",
            "nullable": true
          },
          "build": {
            "description": "Required for a passing `BUILD` job.",
            "$ref": "#/components/schemas/BuildOutputs",
            "nullable": true
          },
          "evidence_graph": {
            "$ref": "#/components/schemas/EvidenceGraph",
            "nullable": true
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionInfo"
          },
          "gates": {
            "description": "Gate results for gates owned by this job kind.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GateResult"
            }
          },
          "log_excerpt": {
            "description": "Short plain-text log excerpt shown publicly (sanitized, bounded to 16 KiB by the server). Must not contain held-out data.",
            "default": null,
            "type": "string",
            "nullable": true
          },
          "manifest": {
            "description": "Required for a passing `VALIDATE` job: the parsed, validated manifest.",
            "$ref": "#/components/schemas/CandidateManifest",
            "nullable": true
          }
        }
      },
      "JobSpec": {
        "description": "A leased job specification, tagged by kind.",
        "oneOf": [
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/ValidateJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "VALIDATE"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/BuildJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "BUILD"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/FormalCheckJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "FORMAL_CHECK"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/ExecJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "CONFORMANCE"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/ExecJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "ADVERSARIAL"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "job",
              "kind"
            ],
            "properties": {
              "job": {
                "$ref": "#/components/schemas/ExecJob"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "BENCHMARK"
                ]
              }
            }
          }
        ]
      },
...
      "LeaseRequest": {
        "type": "object",
        "properties": {
          "kinds": {
            "description": "Job kinds this worker can run (empty = any).",
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobKind"
            }
          },
          "lease_seconds": {
            "description": "Requested lease duration (server clamps to its configured bounds).",
            "default": null,
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0,
            "nullable": true
          }
        }
      },
      "LeasedJob": {
        "type": "object",
        "required": [
          "attempt",
          "job_id",
          "kind",
          "lease_id",
          "lease_until",
          "max_attempts",
          "protocol",
          "run_id",
          "spec",
          "submission_id"
        ],
        "properties": {
          "attempt": {
            "description": "1-based attempt number of this lease.",
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          },
          "job_id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "lease_id": {
            "description": "Fencing token: must be presented on heartbeat/complete/fail. A new lease (e.g. after expiry) invalidates older lease ids.",
            "type": "string"
          },
          "lease_until": {
            "type": "string"
          },
          "max_attempts": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          },
          "protocol": {
            "type": "string"
          },
          "run_id": {
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/JobSpec"
          },
          "submission_id": {
            "type": "string"
          }
        }
      },
...
      "ValidateJob": {
        "type": "object",
        "required": [
          "challenge",
          "ctx"
        ],
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "ctx": {
            "$ref": "#/components/schemas/JobContext"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "agentToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "agent token"
      },
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "admin token (separate table)"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "worker token (worker API only)"
      }
    }
  }
}